Patch 4a68b9

+++ src/tankgame/gameobject/Tank.java
@@ -37,7 +37,7 @@ public class Tank extends GameMovableObject {
         return false;
     }
 
-    public double getHealth() {
+    public int getHealth() {
         return this.health;
     }
